refactor(SystemsTable): Improve integrate-ability and customising
