+++ src/routes/Systems/components/SystemsTable/constants.js
@@ -1,6 +1,17 @@
-export const DEFAULT_OPTIONS = {
-  perPage: 50,
-  onSelect: () => true,
-};
+// TODO Move serialisers.js within SystemsTable directory
+import {
+  filtersSerialiser,
+  sortSerialiser,
+  paginationSerialiser,
+} from '../../serialisers';
 
-export const MAX_PER_PAGE = 100;
+export const DEFAULT_OPTIONS = {
+  // FIXME: remove debug
+  debug: true,
+  perPage: 50,
+  serialisers: {
+    pagination: paginationSerialiser,
+    sort: sortSerialiser,
+    filters: filtersSerialiser,
+  },
+};
